+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_2_9_2021_2_37_PM__Default_Changelist_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_2_9_2021_2_37_PM_[Default_Changelist]" date="1612868873023" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_2_9_2021_2_37_PM_[Default_Changelist]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 2/9/2021 2:37 PM [Default Changelist]" />
+</changelist>
