--- app/src/main/java/com/nima/tmdb/database/DatabaseConnection.kt
@@ -1,6 +1,0 @@
-package com.nima.tmdb.database
-
-interface DatabaseConnection<T>{
-    fun save(T : T)
-    fun load()
-}
