+++ app/src/main/java/com/nima/tmdb/utils/Constants.kt
@@ -15,6 +15,7 @@ object Constants {
     const val DEFAULT_REGION = ""
     const val DEFAULT_PAGE = 1
     const val DEFAULT_ADULT = false
+    const val DATABASE_NAME = "TMDB_DATABASE"
     val DEFAULT_MOVIE_LIST_NAME = arrayOf("life", "toy", "war", "nature", "nina", "big", "peace", "rain", "murder")
     var DEFAULT_IMAGE_REQUEST = RequestOptions().placeholder(R.drawable.ic_launcher_background)
     var DEFAULT_IMAGE = R.drawable.ic_launcher_background.toString()
